requirements.txt
apache-airflow>=2.0.0
openapi-schema-validator<0.2.0
python-gnupg<=0.4.8
